implemented desktop sessions
